Add select with array
