+++ src/store/reducers/books.js
@@ -1,2 +1,15 @@
-//define actionTypes and set a seitch case withthe
-//acion types and return state
+const reducer = (state, action) => {
+  const newState = { ...state };
+  switch (action.type) {
+    case 'ADD_BOOK':
+      break;
+    case 'DELETE_BOOK':
+      break;
+    case 'MODIFY_BOOK':
+      break;
+    default:
+      return newState;
+  }
+  return newState;
+};
+export default reducer;
